Style & autoformat: 4 space soft tabs.

+++ src/org/wildstang/framework/timer/StopWatch.java
@@ -1,65 +1,57 @@
 package org.wildstang.framework.timer;
 
-public class StopWatch
-{
-   private double totalTimeInSec;
-   private long startTimeInNanoSec;
-   private long totalTimeInNanoSec;
-   private boolean timerRunning;
+public class StopWatch {
+    private double totalTimeInSec;
+    private long startTimeInNanoSec;
+    private long totalTimeInNanoSec;
+    private boolean timerRunning;
 
-   public StopWatch()
-   {
-      // Constructor/Initialization Method
-      this.Reset();
-   }
+    public StopWatch() {
+        // Constructor/Initialization Method
+        this.Reset();
+    }
 
-   public void Reset()
-   {
-      startTimeInNanoSec = 0;
-      totalTimeInNanoSec = 0;
-      totalTimeInSec = 0.0;
-      timerRunning = false;
-   }
+    public void Reset() {
+        startTimeInNanoSec = 0;
+        totalTimeInNanoSec = 0;
+        totalTimeInSec = 0.0;
+        timerRunning = false;
+    }
 
-   public void Start()
-   {
-      startTimeInNanoSec = System.nanoTime();
-      timerRunning = true;
-   }
+    public void Start() {
+        startTimeInNanoSec = System.nanoTime();
+        timerRunning = true;
+    }
 
-   public void Stop()
-   {
-      if (timerRunning == true)
-      {
-         long currentTimeInNanoSec = System.nanoTime();
-         long updateTimeInNanoSec;
+    public void Stop() {
+        if (timerRunning == true) {
+            long currentTimeInNanoSec = System.nanoTime();
+            long updateTimeInNanoSec;
 
-         updateTimeInNanoSec = currentTimeInNanoSec - startTimeInNanoSec;
+            updateTimeInNanoSec = currentTimeInNanoSec - startTimeInNanoSec;
 
-         totalTimeInNanoSec += updateTimeInNanoSec;
+            totalTimeInNanoSec += updateTimeInNanoSec;
 
-         startTimeInNanoSec = currentTimeInNanoSec;
+            startTimeInNanoSec = currentTimeInNanoSec;
 
-         timerRunning = false;
-      }
-   }
+            timerRunning = false;
+        }
+    }
 
-   public double GetTimeInSec()
-   {
-      if (timerRunning == true)
-      {
-         long currentTimeInNanoSec = System.nanoTime();
-         long updateTimeInNanoSec;
+    public double GetTimeInSec() {
+        if (timerRunning == true) {
+            long currentTimeInNanoSec = System.nanoTime();
+            long updateTimeInNanoSec;
 
-         updateTimeInNanoSec = currentTimeInNanoSec - startTimeInNanoSec;
+            updateTimeInNanoSec = currentTimeInNanoSec - startTimeInNanoSec;
 
-         totalTimeInNanoSec += updateTimeInNanoSec;
+            totalTimeInNanoSec += updateTimeInNanoSec;
 
-         startTimeInNanoSec = currentTimeInNanoSec;
-      }
+            startTimeInNanoSec = currentTimeInNanoSec;
+        }
 
-      totalTimeInSec = (double) totalTimeInNanoSec / 1000000000.0;
+        totalTimeInSec = totalTimeInNanoSec / 1000000000.0;
 
-      return totalTimeInSec;
-   }
+        return totalTimeInSec;
+    }
 }
